Struct fwdlist::List

pub struct List<T> {
    // some fields omitted
}

A simply linked list.

Methods

impl<T> List<T>

fn new() -> List<T>

A new empty list.

fn len(&self) -> usize

The size of the list in O(1).

fn is_empty(&self) -> bool

Returns true if list is empty in O(1);

fn push_front(&mut self, v: T)

Push a new element at the front of the list in O(1). Cannot fails, only panic!/OOM on memory exhaustion.

fn pop_front(&mut self) -> Option<T>

Pop a element from the front of the list in O(1). Returns None if the list is empty.

fn push_back(&mut self, v: T)

Push an element at the end of the list in O(n). Cannot fails, only panic!/OOM on memory exhaustion.

fn pop_back(&mut self) -> Option<T>

Pop an element from the end of the list in O(n). Returns None if the list is empty.

fn clear(&mut self)

Clear the list in O(n).

impl<T> List<T>

Some accessors to front/back elements.

fn front(&self) -> Option<&T>

Returns a reference to the first element in the list.

fn front_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the first element in the list.

fn back(&self) -> Option<&T>

Returns a reference to the last element in the list.

fn back_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the last element in the list.

impl<T> List<T>

Extra operations on the list - Unstable API.

fn append(&mut self, other: &mut List<T>)

Moves all elements from other to the end of the list in O(self.len());

fn split_off(&mut self, at: usize) -> List<T>

Splits the list into two at the given index in O(at).

Returns everything after the given index, including the index.

if at == self.len(), returns an empty list in O(1).

Panics if at > self.len().

impl<T> List<T>

fn iter(&self) -> ListIter<T>

Returns an iterator over the list yielding read-only references.

impl<T> List<T>

fn iter_mut(&mut self) -> ListIterMut<T>

Returns an iterator over the list yielding mutable references.

impl<T> List<T>

fn cursor(&mut self) -> Cursor<T>

Return a cursor at the beginning of the list (before the first node).